extract internal components from bingo assistant
